Run clang-tidy ourselves, and fix what it found
The code-quality job had never checked anything. Its `make tidy-check`
configures the whole of DuckDB, which needs the dependencies vcpkg.json
declares, and the reusable workflow does not set vcpkg up: the configure died
on `find_package(tinyxml2)` before clang-tidy started. That is what the red
mark was. The workflow takes an `extra_toolchains` input that looks like the
place to ask for vcpkg, but nothing in the file reads it. And its file filter,
`<project>/src/.*/`, matches nothing in a flat `src/*.cpp` layout, so even a
successful configure would have checked no files -- which is worth knowing,
because it means the same job is decorative in every extension laid out this
way.

scripts/tidy.sh runs the checker directly with DuckDB's own .clang-tidy, over
this extension's sources, with a header filter pointed at our headers rather
than DuckDB's. No compile database, no DuckDB build, about a minute.

It found around thirty things, all real, none subtle: multiplications done in
int and widened on assignment to int64_t, `long` where int64_t was meant,
`push_back` in a loop without a reserve, a const local that could not be moved
out of on return, and a single-argument constructor without explicit. Three
places are suppressed with a reason rather than changed: libcurl writes a
response code through a `long*`, execv wants `char *const[]`, and `environ` is
POSIX's own declaration. Vendored base64.cpp is excluded -- rewriting it would
turn the next update from a copy into a merge.

+++ src/metadata_fetch.cpp
@@ -286,7 +286,7 @@ std::vector<ColumnInfo> FetchColumnsOfTables(FusionTransport &transport, const R
 	std::unordered_set<std::string> seen;
 	for (size_t start = 0; start < ids.size(); start += metadata::COLUMN_BATCH_SIZE) {
 		const auto end = std::min(start + metadata::COLUMN_BATCH_SIZE, ids.size());
-		const std::vector<std::string> batch(ids.begin() + static_cast<long>(start), ids.begin() + static_cast<long>(end));
+		const std::vector<std::string> batch(ids.begin() + static_cast<std::ptrdiff_t>(start), ids.begin() + static_cast<std::ptrdiff_t>(end));
 
 		for (const auto &row : QueryPaged(transport, context, metadata::ColumnsByTableIds(batch))) {
 			ColumnInfo column;
